Added working files without virtual environment and dependencies
Only the working and resulting project files are posted in this repository. The folder with the virtual environment in which the work was carried out is not included in the repository. There is also a missing file with dependencies (for example, requirements.txt or Vip file).

To reproduce the project, please create a new virtual environment and install the necessary dependencies manually.